+++ HealthBarColor/Modules/Fonts/Focus.lua
@@ -13,7 +13,7 @@ local Font_Focus = HealthBarColor:NewModule("Font_Focus")
 local Focus = HealthBarColor:GetUnit("Focus")
 local ToF = HealthBarColor:GetUnit("ToF")
 local Media = LibStub("LibSharedMedia-3.0")
-local name_font, healthbar_font, powerbar_font, name_outline, healthbar_outline, powerbar_outline
+local name_font, healthbar_font, powerbar_font, name_outline, healthbar_outline, powerbar_outline, name_size, healthbar_size, powerbar_size
 local name_color, healthbar_color, powerbar_color = {r=0,g=0,b=0}, {r=0,g=0,b=0}, {r=0,g=0,b=0}
 
 function Font_Focus:OnEnable()
@@ -23,14 +23,17 @@ function Font_Focus:OnEnable()
     local selected = HealthBarColor.db.profile.Fonts.General.name_outline
     name_outline = selected == 1 and "OUTLINE"  or selected == 2 and "THICKOUTLINE"  or selected == 3 and "MONOCHROME"  or selected == 4 and "NONE" 
     name_font = Media:Fetch("font",HealthBarColor.db.profile.Fonts.General.name) 
+    name_size = HealthBarColor.db.profile.Fonts.General.name_size
     --healthbar
     selected = HealthBarColor.db.profile.Fonts.General.healthbar_outline
     healthbar_outline = selected == 1 and "OUTLINE"  or selected == 2 and "THICKOUTLINE"  or selected == 3 and "MONOCHROME"  or selected == 4 and "NONE" 
     healthbar_font = Media:Fetch("font",HealthBarColor.db.profile.Fonts.General.healthbar) 
+    healthbar_size = HealthBarColor.db.profile.Fonts.General.healthbar_size
     --powerbar
     selected = HealthBarColor.db.profile.Fonts.General.powerbar_outline
     powerbar_outline = selected == 1 and "OUTLINE"  or selected == 2 and "THICKOUTLINE"  or selected == 3 and "MONOCHROME"  or selected == 4 and "NONE"
     powerbar_font = Media:Fetch("font",HealthBarColor.db.profile.Fonts.General.powerbar) 
+    powerbar_size = HealthBarColor.db.profile.Fonts.General.powerbar_size
     self:SetFonts()
     --COLOR
     --name
@@ -174,12 +177,12 @@ function Font_Focus:OnDisable()
 end
 
 function Font_Focus:SetFonts()
-    Focus.Name:SetFont(name_font, 10, name_outline)
-    ToF.Name:SetFont(name_font, 10, name_outline)
+    Focus.Name:SetFont(name_font, name_size, name_outline)
+    ToF.Name:SetFont(name_font, name_size, name_outline)
     for _,text in pairs (Focus.HealthText) do
-        text:SetFont(healthbar_font, 10,healthbar_outline)
+        text:SetFont(healthbar_font, healthbar_size, healthbar_outline)
     end
     for _,text in pairs (Focus.PowerText) do
-        text:SetFont(powerbar_font, 10,powerbar_outline)
+        text:SetFont(powerbar_font, powerbar_size, powerbar_outline)
     end
 end