+++ HealthBarColor/Modules/Fonts/Player.lua
@@ -9,7 +9,7 @@ local HealthBarColor = addonTable.HealthBarColor
 local Font_Player = HealthBarColor:NewModule("Font_Player")
 local Player = HealthBarColor:GetUnit("Player")
 local Media = LibStub("LibSharedMedia-3.0")
-local name_font, healthbar_font, powerbar_font, name_outline, healthbar_outline, powerbar_outline
+local name_font, healthbar_font, powerbar_font, name_outline, healthbar_outline, powerbar_outline, name_size, healthbar_size, powerbar_size
 local name_color, healthbar_color, powerbar_color = {r=0,g=0,b=0}, {r=0,g=0,b=0}, {r=0,g=0,b=0}
 
 
@@ -20,14 +20,17 @@ function Font_Player:OnEnable()
     local selected = HealthBarColor.db.profile.Fonts.General.name_outline
     name_outline = selected == 1 and "OUTLINE"  or selected == 2 and "THICKOUTLINE"  or selected == 3 and "MONOCHROME"  or selected == 4 and "NONE" 
     name_font = Media:Fetch("font",HealthBarColor.db.profile.Fonts.General.name) 
+    name_size = HealthBarColor.db.profile.Fonts.General.name_size
     --healthbar
     selected = HealthBarColor.db.profile.Fonts.General.healthbar_outline
     healthbar_outline = selected == 1 and "OUTLINE"  or selected == 2 and "THICKOUTLINE"  or selected == 3 and "MONOCHROME"  or selected == 4 and "NONE" 
     healthbar_font = Media:Fetch("font",HealthBarColor.db.profile.Fonts.General.healthbar) 
+    healthbar_size = HealthBarColor.db.profile.Fonts.General.healthbar_size
     --powerbar
     selected = HealthBarColor.db.profile.Fonts.General.powerbar_outline
     powerbar_outline = selected == 1 and "OUTLINE"  or selected == 2 and "THICKOUTLINE"  or selected == 3 and "MONOCHROME"  or selected == 4 and "NONE"
     powerbar_font = Media:Fetch("font",HealthBarColor.db.profile.Fonts.General.powerbar) 
+    powerbar_size = HealthBarColor.db.profile.Fonts.General.powerbar_size
     --COLOR
     --name
     if HealthBarColor.db.profile.Fonts.Player.name_choice == 1 then
@@ -47,7 +50,7 @@ function Font_Player:OnEnable()
     else
         powerbar_color = HealthBarColor.db.profile.Fonts.Player.powerbar_color
     end 
-    self:Setup()
+    self:SetFonts()
 end
 
 function Font_Player:OnDisable()
@@ -63,15 +66,15 @@ function Font_Player:OnDisable()
     end
 end
 
-function Font_Player:Setup()
-    Player.Name:SetFont(name_font, 10, name_outline)
+function Font_Player:SetFonts()
+    Player.Name:SetFont(name_font, name_size, name_outline)
     Player.Name:SetTextColor(name_color.r,name_color.g,name_color.b)
     for _,text in pairs (Player.HealthText) do
-        text:SetFont(healthbar_font, 10,healthbar_outline)
+        text:SetFont(healthbar_font, healthbar_size, healthbar_outline)
         text:SetTextColor(healthbar_color.r,healthbar_color.g,healthbar_color.b)
     end
     for _,text in pairs (Player.PowerText) do
-        text:SetFont(powerbar_font, 10,powerbar_outline)
+        text:SetFont(powerbar_font, powerbar_size, powerbar_outline)
         text:SetTextColor(powerbar_color.r,powerbar_color.g,powerbar_color.b)
     end
 end